# backend/src/routers/auth.py
import logging
from datetime import datetime, timedelta, timezone

from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy import delete, select
from sqlalchemy.ext.asyncio import AsyncSession

from ..auth import (
    generate_email_code,
    hash_email_code,
    hash_password,
    hash_refresh_token,
    verify_password,
)
from ..config import EMAIL_CODE_EXPIRE_MINUTES, MAX_CODE_SENDS_PER_10_MINUTES
from ..database import get_db
from ..email_service import EmailService
from ..models import EmailCodesStorage, RefreshToken, User
from ..schemas import (
    AuthResponse,
    ForgotPassword,
    ForgotPasswordVerify,
    LoginRequest,
    MessageResponse,
    RefreshRequest,
    SendCodeRequest,
    UserResponse,
    VerifyCodeRequest,
)
from ..services.auth_tokens import issue_tokens, revoke_user_sessions
from ..utils import (
    DatabaseRateLimiter,
    PasswordValidationError,
    validate_password,
    with_rate_limit,
)

router = APIRouter(tags=["Auth"])
post = with_rate_limit(router.post, DatabaseRateLimiter(100, 1))
logger = logging.getLogger(__name__)


def _utcnow_naive() -> datetime:
    return datetime.now(timezone.utc).replace(tzinfo=None)


@post("/auth/send-code", response_model=MessageResponse)
async def send_code(body: SendCodeRequest, db: AsyncSession = Depends(get_db)):
    """Send verification code to email"""
    # Clean up expired codes
    await _cleanup_expired_codes(db)

    # Check rate limiting
    codes = await db.execute(
        select(EmailCodesStorage).where(
            EmailCodesStorage.email == body.email,
            EmailCodesStorage.created_at
            > (_utcnow_naive() - timedelta(minutes=EMAIL_CODE_EXPIRE_MINUTES)),
        )
    )
    recent_codes = codes.scalars().all()
    if len(recent_codes) >= MAX_CODE_SENDS_PER_10_MINUTES:
        raise HTTPException(
            status_code=status.HTTP_429_TOO_MANY_REQUESTS,
            detail="Too many verification codes sent. Please wait 10 minutes.",
        )

    code = generate_email_code()
    storage = EmailCodesStorage(
        email=str(body.email),
        code_hash=hash_email_code(str(body.email), code),
    )
    db.add(storage)
    await db.commit()

    # Send email
    await EmailService.send_code(str(body.email), code)

    return {"message": "Verification code sent"}


@post("/auth/verify-code", response_model=AuthResponse)
async def verify_code(body: VerifyCodeRequest, db: AsyncSession = Depends(get_db)):
    """Verify code and return tokens (for login/registration)"""
    # Clean up expired codes
    await _cleanup_expired_codes(db)

    result = await db.execute(
        select(EmailCodesStorage).where(
            EmailCodesStorage.email == body.email,
            EmailCodesStorage.code_hash == hash_email_code(str(body.email), body.code),
            EmailCodesStorage.created_at
            > _utcnow_naive() - timedelta(minutes=EMAIL_CODE_EXPIRE_MINUTES),
        )
    )
    code_storage = result.scalar_one_or_none()

    if not code_storage:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid or expired verification code",
        )

    # Delete used code
    await db.delete(code_storage)

    # Find or create user
    result = await db.execute(select(User).where(User.email == body.email))
    user = result.scalar_one_or_none()

    if user is None:
        # Registration: create a new user
        if not body.password:
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail="Password is required for registration",
            )
        # Validate password
        try:
            validate_password(body.password)
        except PasswordValidationError as e:
            raise HTTPException(
                status_code=status.HTTP_400_BAD_REQUEST,
                detail=str(e),
            ) from e
        user = User(
            email=body.email,
            password_hash=hash_password(body.password),
        )
        db.add(user)
        await db.flush()
    else:
        # Update password if provided (for password reset flow)
        if body.password:
            # Validate password
            try:
                validate_password(body.password)
            except PasswordValidationError as e:
                raise HTTPException(
                    status_code=status.HTTP_400_BAD_REQUEST,
                    detail=str(e),
                ) from e
            user.password_hash = hash_password(body.password)
            await revoke_user_sessions(db, user)

    access_token, refresh_token_str = await issue_tokens(db, user)

    return AuthResponse(
        accessToken=access_token,
        refreshToken=refresh_token_str,
        user=UserResponse(
            id=user.id,
            email=user.email,
            premium=user.premium or False,
            subscriptionExpires=user.subscription_expires,
        ),
    )


@post("/auth/login", response_model=AuthResponse)
async def login(body: LoginRequest, db: AsyncSession = Depends(get_db)):
    """Legacy login endpoint - kept for backward compatibility"""
    result = await db.execute(select(User).where(User.email == body.email))
    user = result.scalar_one_or_none()

    if not user or not verify_password(body.password, user.password_hash):
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid email or password",
        )

    access_token, refresh_token_str = await issue_tokens(db, user)

    return AuthResponse(
        accessToken=access_token,
        refreshToken=refresh_token_str,
        user=UserResponse(
            id=user.id,
            email=user.email,
            premium=user.premium or False,
            subscriptionExpires=user.subscription_expires,
        ),
    )


@post("/auth/forgot-password/send-code", response_model=MessageResponse)
async def forgot_password_send_code(
    body: ForgotPassword,
    db: AsyncSession = Depends(get_db),
):
    """Send password reset code"""
    # Clean up expired codes
    await _cleanup_expired_codes(db)

    # Check if user exists
    result = await db.execute(select(User).where(User.email == body.email))
    user = result.scalar_one_or_none()

    if not user:
        # Don't reveal if email exists or not
        logger.info("Password reset requested")
        return {"message": "If the email exists, a reset code has been sent"}

    # Check rate limiting — count codes created in the last N minutes
    result = await db.execute(
        select(EmailCodesStorage).where(
            EmailCodesStorage.email == body.email,
            EmailCodesStorage.created_at
            > _utcnow_naive() - timedelta(minutes=EMAIL_CODE_EXPIRE_MINUTES),
        )
    )
    recent_codes = result.scalars().all()
    if len(recent_codes) >= MAX_CODE_SENDS_PER_10_MINUTES:
        raise HTTPException(
            status_code=status.HTTP_429_TOO_MANY_REQUESTS,
            detail="Too many reset codes sent. Please wait 10 minutes.",
        )

    # Generate and store code
    code = generate_email_code()
    storage = EmailCodesStorage(
        email=str(body.email),
        code_hash=hash_email_code(str(body.email), code),
    )
    db.add(storage)
    await db.commit()

    # Send email
    await EmailService.send_code(
        body.email,
        code,
        subject="Сброс пароля",
    )

    return {"message": "If the email exists, a reset code has been sent"}


@post("/auth/forgot-password/verify-code", response_model=MessageResponse)
async def forgot_password_verify_code(
    body: ForgotPasswordVerify, db: AsyncSession = Depends(get_db)
):
    """Verify reset code and set new password"""
    # Clean up expired codes
    await _cleanup_expired_codes(db)

    # Find a valid (non-expired) code
    result = await db.execute(
        select(EmailCodesStorage).where(
            EmailCodesStorage.email == body.email,
            EmailCodesStorage.code_hash == hash_email_code(str(body.email), body.code),
            EmailCodesStorage.created_at
            > _utcnow_naive() - timedelta(minutes=EMAIL_CODE_EXPIRE_MINUTES),
        )
    )
    code_storage = result.scalar_one_or_none()

    if not code_storage:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid or expired reset code",
        )

    # Delete used code
    await db.delete(code_storage)

    # Validate new password
    try:
        validate_password(body.new_password)
    except PasswordValidationError as e:
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail=str(e),
        ) from e

    # Find user and update password
    result = await db.execute(select(User).where(User.email == body.email))
    user = result.scalar_one_or_none()

    if not user:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="User not found",
        )

    user.password_hash = hash_password(body.new_password)
    await revoke_user_sessions(db, user)

    return {"message": "Password reset successful"}


@post("/auth/refresh", response_model=AuthResponse)
async def refresh(body: RefreshRequest, db: AsyncSession = Depends(get_db)):
    result = await db.execute(
        select(RefreshToken).where(RefreshToken.token_hash == hash_refresh_token(body.refreshToken))
    )
    rt = result.scalar_one_or_none()
    if not rt:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid refresh token",
        )

    if rt.expires_at.replace(tzinfo=None) < _utcnow_naive():
        await db.delete(rt)
        await db.commit()
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Refresh token expired",
        )

    user_result = await db.execute(select(User).where(User.id == rt.user_id))
    user = user_result.scalar_one_or_none()
    if not user:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="User not found",
        )

    # Delete old refresh token
    await db.delete(rt)

    access_token, refresh_token_str = await issue_tokens(db, user)

    return AuthResponse(
        accessToken=access_token,
        refreshToken=refresh_token_str,
        user=UserResponse(
            id=user.id,
            email=user.email,
            premium=user.premium or False,
            subscriptionExpires=user.subscription_expires,
        ),
    )


async def _cleanup_expired_codes(db: AsyncSession) -> None:
    """Remove expired email verification codes"""
    cutoff = _utcnow_naive() - timedelta(minutes=EMAIL_CODE_EXPIRE_MINUTES)
    await db.execute(
        delete(EmailCodesStorage)
        .where(EmailCodesStorage.created_at < cutoff)
        .execution_options(synchronize_session=False)
    )
    await db.commit()